Ensure .emacs.d is a git repository
The current copy operation omits hidden files and folders which means that `.git` is not copied and `.emacs.d` is not a git respository. This makes it much more difficult to update emacs-live later.

This change ensures that the entire contents of `$tmp_dir/overtone-emacs-live` is copied to `~/.emacs.d`, including `.git`.