Fixes #153: Ediff save bug.
Some buffers are dedicated for certain processes and should never be
asked for filenames for saving (e.g. certain windows for ediff, probably
org-mode calendar windows, etc.).

live-mp-new-basic-save-buffer was breaking save/close functionality of
ediff. Now that there is a check that frame-selected-window is not a
dedicated window before changing to the current buffer, these problems
are gone (Fixes #153).