Set correct org-odt-data-dir
The variable org-odt-data-dir is defined in org-loaddefs.el and
org-version.el, which are created when running make or make autoloads for
org-mode.

Emacs-live bundles its own org-mode, versus using an org-mode from the host
Emacs application.  The default for org-odt-data-dir is a path to
an org-mode bundled with Emacs installed in /usr/share, but should point
to the org-mode included with Emacs-live.

Configuration for org-mode has been updated to set org-odt-data-dir based on
the location of Emacs-live after loading org-loaddefs.el and org-version.el.